# Conflicts:
#	README.md
#	pom.xml
#	src/main/java/com/backend/Application.java
#	src/main/java/com/backend/dto/PartnerDTO.java
#	src/main/java/com/backend/repository/IPartnerRepository.java
#	src/main/java/com/backend/util/DataUtil.java
#	src/main/resources/application.properties
#	src/main/resources/log4j2.xml
#	src/test/java/com/backend/ApplicationTests.java
#	src/test/java/com/backend/DemoApplicationTests.java
#	src/test/java/com/example/demo/DemoApplicationTests.java

+++ src/main/java/com/backend/repository/IPartnerRepository.java
@@ -3,5 +3,5 @@ package com.backend.repository;
 import com.backend.dto.PartnerDTO;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IPartnerRepository extends CrudRepository<PartnerDTO, Long> {
+public interface IPartnerRepository extends CrudRepository<PartnerDTO, Integer> {
 }
